+++ src/braket/aws/aws_device.py
@@ -22,6 +22,7 @@ from typing import Dict, List, Optional, Union
 from botocore.errorfactory import ClientError
 from networkx import DiGraph, complete_graph, from_edgelist
 
+from braket.ahs.analog_hamiltonian_simulation import AnalogHamiltonianSimulation
 from braket.annealing.problem import Problem
 from braket.aws.aws_quantum_task import AwsQuantumTask
 from braket.aws.aws_quantum_task_batch import AwsQuantumTaskBatch
@@ -90,7 +91,12 @@ class AwsDevice(Device):
     def run(
         self,
         task_specification: Union[
-            Circuit, Problem, OpenQasmProgram, BlackbirdProgram, PulseSequence
+            Circuit,
+            Problem,
+            OpenQasmProgram,
+            BlackbirdProgram,
+            PulseSequence,
+            AnalogHamiltonianSimulation,
         ],
         s3_destination_folder: Optional[AwsSession.S3DestinationFolder] = None,
         shots: Optional[int] = None,
@@ -104,7 +110,7 @@ class AwsDevice(Device):
         annealing problem.
 
         Args:
-            task_specification (Union[Circuit, Problem, OpenQasmProgram, BlackbirdProgram, PulseSequence]): # noqa
+            task_specification (Union[Circuit, Problem, OpenQasmProgram, BlackbirdProgram, PulseSequence, AnalogHamiltonianSimulation]): # noqa
                 Specification of task (circuit or annealing problem or program) to run on device.
             s3_destination_folder (Optional[S3DestinationFolder]): The S3 location to
                 save the task's results to. Default is `<default_bucket>/tasks` if evoked
@@ -170,7 +176,14 @@ class AwsDevice(Device):
     def run_batch(
         self,
         task_specifications: List[
-            Union[Circuit, Problem, OpenQasmProgram, BlackbirdProgram, PulseSequence]
+            Union[
+                Circuit,
+                Problem,
+                OpenQasmProgram,
+                BlackbirdProgram,
+                PulseSequence,
+                AnalogHamiltonianSimulation,
+            ]
         ],
         s3_destination_folder: Optional[AwsSession.S3DestinationFolder] = None,
         shots: Optional[int] = None,
@@ -184,7 +197,7 @@ class AwsDevice(Device):
         """Executes a batch of tasks in parallel
 
         Args:
-            task_specifications (List[Union[Circuit, Problem, OpenQasmProgram, BlackbirdProgram, PulseSequence]]): # noqa
+            task_specifications (List[Union[Circuit, Problem, OpenQasmProgram, BlackbirdProgram, PulseSequence, AnalogHamiltonianSimulation]]): # noqa
                 List of  circuits or annealing problems to run on device.
             s3_destination_folder (Optional[S3DestinationFolder]): The S3 location to
                 save the tasks' results to. Default is `<default_bucket>/tasks` if evoked